Add detectors to analyze SQL warnings and refactor structure
Introduce `FunctionInvalidatesIndexDetector` and integrate new detectors into `ExplainAnalyzer` for improved SQL warning detection. Refactor detectors into a dedicated namespace (`Koriym\SqlQuality\Detector`) and implement `DetectorInterface` for consistency. These updates enhance modularity and extendability of the codebase.
